Moved DescriptionModificationAllWithParentChildRelationshipPrecondition to ManipulationPreconditions.
