+++ sdccc/src/main/java/com/draeger/medical/sdccc/manipulation/precondition/impl/ConditionalPreconditions.java
@@ -187,48 +187,6 @@ public class ConditionalPreconditions {
         return ResponseTypes.Result.RESULT_SUCCESS.equals(manipulationResult);
     }
 
-    private static boolean descriptionUpdateWithParentChildRelationshipManipulation(
-            final Injector injector, final Logger log) {
-        final Pair<String, String> descriptorHandles = findFirstDescriptorHandleWithChildren(injector, log);
-        if (descriptorHandles == null) {
-            log.error("Could not trigger a descriptor update with parent child relationship as there are no "
-                    + "descriptors with children in the mdib.");
-            return false;
-        }
-        final String parentDescriptorHandle = descriptorHandles.getLeft();
-        final String childDescriptorHandle = descriptorHandles.getRight();
-        final var manipulations = injector.getInstance(Manipulations.class);
-
-        final ResponseTypes.Result manipulationResult =
-                manipulations.triggerDescriptorUpdate(List.of(childDescriptorHandle, parentDescriptorHandle));
-
-        return ResponseTypes.Result.RESULT_SUCCESS.equals(manipulationResult);
-    }
-
-    private static Pair<String, String> findFirstDescriptorHandleWithChildren(
-            final Injector injector, final Logger log) {
-        final var testClient = injector.getInstance(TestClient.class);
-
-        final var remoteDevice = testClient.getSdcRemoteDevice();
-        if (remoteDevice == null) {
-            log.error("remote device could not be accessed, likely due to a disconnect");
-            return null;
-        }
-        final var mdibAccess = remoteDevice.getMdibAccess();
-
-        final Collection<MdibEntity> entities = mdibAccess.findEntitiesByType(AbstractDescriptor.class);
-        for (MdibEntity entity : entities) {
-            if (entity.getDescriptor() instanceof MdsDescriptor) {
-                continue; // we are not interested in MdsDescriptors as many Devices do not support modifying them.
-            }
-            final List<String> children = entity.getChildren();
-            if (!children.isEmpty()) {
-                return Pair.of(entity.getHandle(), children.get(0));
-            }
-        }
-        return null;
-    }
-
     private static boolean triggerReportPreconditionCheck(
             final Injector injector, final Logger log, final QName... reportType) throws PreconditionException {
         final var messageStorage = injector.getInstance(MessageStorage.class);
@@ -730,41 +688,6 @@ public class ConditionalPreconditions {
          */
         static boolean manipulation(final Injector injector) {
             return descriptionModificationManipulation(injector, LOG);
-        }
-    }
-
-    /**
-     * Precondition which triggers all possible description changes with parent-child relationships.
-     */
-    public static class DescriptionModificationAllWithParentChildRelationshipPrecondition extends SimplePrecondition {
-
-        private static final Logger LOG =
-                LogManager.getLogger(DescriptionModificationAllWithParentChildRelationshipPrecondition.class);
-
-        /**
-         * Creates a description modification all with parent child relationship precondition check.
-         */
-        public DescriptionModificationAllWithParentChildRelationshipPrecondition() {
-            super(
-                    DescriptionModificationAllWithParentChildRelationshipPrecondition::preconditionCheck,
-                    DescriptionModificationAllWithParentChildRelationshipPrecondition::manipulation);
-        }
-
-        static boolean preconditionCheck(final Injector injector) {
-            return false; // always trigger description changes
-        }
-
-        /**
-         * Performs the removal, reinsertion and update of all modifiable descriptors in the mdib to trigger reports.
-         *
-         * @param injector to analyze mdib on
-         * @return true if successful, false otherwise
-         * @throws PreconditionException on errors
-         */
-        static boolean manipulation(final Injector injector) {
-            final boolean result1 = descriptionModificationManipulation(injector, LOG);
-            final boolean result2 = descriptionUpdateWithParentChildRelationshipManipulation(injector, LOG);
-            return result1 && result2;
         }
     }
