+++ sdccc/src/test/java/com/draeger/medical/sdccc/manipulation/precondition/impl/ConditionalPreconditionsTest.java
@@ -369,101 +369,6 @@ public class ConditionalPreconditionsTest {
     }
 
     /**
-     * Tests whether DescriptionModificationAllWithParentChildRelationshipPrecondition.preconditionCheck
-     * always returns false.
-     */
-    @Test
-    @DisplayName("DescriptionModificationAllWithParentChildRelationshipPrecondition preconditionCheck always returns"
-            + " false to trigger the manipulation.")
-    public void testDescriptionModificationAllWithParentChildRelationshipPreconditionCheck() {
-        // no messages - does not matter as the preconditionCheck should always return false
-        assertFalse(
-                ConditionalPreconditions.DescriptionModificationAllWithParentChildRelationshipPrecondition
-                        .preconditionCheck(testInjector));
-    }
-
-    /**
-     * Tests whether DescriptionModificationAllWithParentChildRelationshipPrecondition correctly calls manipulation.
-     */
-    @Test
-    @DisplayName("DescriptionModificationAllWithParentChildRelationshipPrecondition correctly calls manipulation")
-    public void testDescriptionModificationAllWithParentChildRelationshipPreconditionManipulation() {
-
-        final var descriptor1Handle = "superHandle";
-        final var descriptor2Handle = "handle;Süper;";
-        final var parentDescriptorHandle = "parentHandle";
-        final var childDescriptorHandle = "childHandle";
-
-        final var presenceMap = new HashMap<>(Map.of(
-                descriptor1Handle, false,
-                descriptor2Handle, true));
-
-        when(mockManipulations.getRemovableDescriptorsOfClass())
-                .thenReturn(List.of(descriptor1Handle, descriptor2Handle));
-
-        when(mockManipulations.insertDescriptor(anyString())).thenAnswer((Answer<ResponseTypes.Result>) invocation -> {
-            presenceMap.put(invocation.getArgument(0), true);
-            return ResponseTypes.Result.RESULT_SUCCESS;
-        });
-        when(mockManipulations.removeDescriptor(anyString())).thenAnswer((Answer<ResponseTypes.Result>) invocation -> {
-            presenceMap.put(invocation.getArgument(0), false);
-            return ResponseTypes.Result.RESULT_SUCCESS;
-        });
-        when(mockManipulations.triggerDescriptorUpdate(anyList())).thenReturn(ResponseTypes.Result.RESULT_SUCCESS);
-        final MdibEntity mockEntity = mock(MdibEntity.class);
-        when(mockEntity.getHandle()).thenReturn(parentDescriptorHandle);
-        when(mockEntity.getChildren()).thenReturn(List.of(childDescriptorHandle));
-        when(testClient.getSdcRemoteDevice().getMdibAccess().findEntitiesByType(any()))
-                .thenReturn(List.of(mockEntity));
-        final MdibEntity mockEntity2 = mock(MdibEntity.class);
-        when(testClient.getSdcRemoteDevice().getMdibAccess().getEntity(any())).thenAnswer(args -> {
-            if (presenceMap.get((String) args.getArgument(0))) {
-                return Optional.of(mockEntity2);
-            } else {
-                return Optional.empty();
-            }
-        });
-
-        assertTrue(
-                ConditionalPreconditions.DescriptionModificationAllWithParentChildRelationshipPrecondition.manipulation(
-                        testInjector));
-
-        final var insertCaptor = ArgumentCaptor.forClass(String.class);
-        final var removeCaptor = ArgumentCaptor.forClass(String.class);
-        final var handleCaptor = ArgumentCaptor.forClass(List.class);
-        verify(mockManipulations, times(1)).getRemovableDescriptorsOfClass();
-        verify(mockManipulations, times(3)).insertDescriptor(insertCaptor.capture());
-        verify(mockManipulations, times(2)).removeDescriptor(removeCaptor.capture());
-        verify(mockManipulations, times(1)).triggerDescriptorUpdate(handleCaptor.capture());
-
-        assertEquals(2, handleCaptor.getAllValues().get(0).size());
-        assertEquals(childDescriptorHandle, handleCaptor.getAllValues().get(0).get(0));
-        assertEquals(parentDescriptorHandle, handleCaptor.getAllValues().get(0).get(1));
-
-        assertEquals(
-                2,
-                insertCaptor.getAllValues().stream()
-                        .filter(descriptor1Handle::equals)
-                        .count());
-        assertEquals(
-                1,
-                insertCaptor.getAllValues().stream()
-                        .filter(descriptor2Handle::equals)
-                        .count());
-
-        assertEquals(
-                1,
-                removeCaptor.getAllValues().stream()
-                        .filter(descriptor1Handle::equals)
-                        .count());
-        assertEquals(
-                1,
-                removeCaptor.getAllValues().stream()
-                        .filter(descriptor2Handle::equals)
-                        .count());
-    }
-
-    /**
      * Tests whether DescriptionModificationCrtPrecondition correctly checks for precondition.
      *
      * @throws PreconditionException on precondition exceptions
